Initial working code for movieListPage
